requirements/requirements_multimodal.txt
addict
bitsandbytes==0.45.0 ; platform_machine == 'x86_64'
clip
decord; sys_platform == 'linux'
diffusers>=0.19.3
einops_exts
imageio
kornia
megatron-energon==5.1.0
nerfacc>=0.5.3
open_clip_torch==2.24.0
PyMCubes
taming-transformers
torchdiffeq
torchsde
trimesh
